Move more functions to _utils
